Rename files to match exports
