+++ packages/theme/src/components/MuiAvatar.stories.tsx
