+++ packages/theme/src/components/MuiAvatar.ts
