+++ packages/theme/src/components/MuiCheckbox.ts
