+++ packages/theme/src/components/MuiDialogActions.ts
