+++ packages/theme/src/components/MuiDialogContent.ts
