+++ packages/theme/src/components/MuiListItemText.ts
