+++ packages/theme/src/components/MuiTabs.ts
