+++ packages/theme/src/components/index.ts
@@ -2,48 +2,48 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-export * from "./autocomplete";
-export * from "./avatar";
-export * from "./badge";
-export * from "./backdrop";
-export * from "./button";
-export * from "./buttonBase";
-export * from "./card";
-export * from "./cardActionArea";
-export * from "./cardContent";
-export * from "./cardHeader";
-export * from "./checkbox";
-export * from "./chip";
-export * from "./cssBaseline";
-export * from "./dialog";
-export * from "./dialogActions";
-export * from "./dialogContent";
-export * from "./dialogTitle";
-export * from "./fab";
-export * from "./filledInput";
-export * from "./formControl";
-export * from "./formLabel";
-export * from "./iconButton";
-export * from "./input";
-export * from "./inputAdornment";
-export * from "./inputBase";
-export * from "./inputLabel";
-export * from "./link";
-export * from "./listItemButton";
-export * from "./listItemText";
-export * from "./listSubheader";
-export * from "./menu";
-export * from "./menuItem";
-export * from "./outlinedInput";
-export * from "./paper";
-export * from "./popover";
-export * from "./radio";
-export * from "./select";
-export * from "./tab";
-export * from "./tableCell";
-export * from "./tabs";
-export * from "./textField";
-export * from "./toggleButton";
-export * from "./toolbar";
-export * from "./tooltip";
-export * from "./typography";
+export * from "./MuiAutocomplete";
+export * from "./MuiAvatar";
+export * from "./MuiBadge";
+export * from "./MuiBackdrop";
+export * from "./MuiButton";
+export * from "./MuiButtonBase";
+export * from "./MuiCard";
+export * from "./MuiCardActionArea";
+export * from "./MuiCardContent";
+export * from "./MuiCardHeader";
+export * from "./MuiCheckbox";
+export * from "./MuiChip";
+export * from "./MuiCssBaseline";
+export * from "./MuiDialog";
+export * from "./MuiDialogActions";
+export * from "./MuiDialogContent";
+export * from "./MuiDialogTitle";
+export * from "./MuiFab";
+export * from "./MuiFilledInput";
+export * from "./MuiFormControl";
+export * from "./MuiFormLabel";
+export * from "./MuiIconButton";
+export * from "./MuiInput";
+export * from "./MuiInputAdornment";
+export * from "./MuiInputBase";
+export * from "./MuiInputLabel";
+export * from "./MuiLink";
+export * from "./MuiListItemButton";
+export * from "./MuiListItemText";
+export * from "./MuiListSubheader";
+export * from "./MuiMenu";
+export * from "./MuiMenuItem";
+export * from "./MuiOutlinedInput";
+export * from "./MuiPaper";
+export * from "./MuiPopover";
+export * from "./MuiRadio";
+export * from "./MuiSelect";
+export * from "./MuiTab";
+export * from "./MuiTableCell";
+export * from "./MuiTabs";
+export * from "./MuiTextField";
+export * from "./MuiToggleButton";
+export * from "./MuiToolbar";
+export * from "./MuiTooltip";
+export * from "./MuiTypography";
